# add CPU feature checks for AVX2