Fixes #189 

The `bloom` package was making the assumption that AVX2 would be available, however it appears that some users are running programs that depend on parquet-go on intel CPUs which do not have AVX2.

The PR adds a check for the AVX2 CPU feature in the assembly functions for amd64. The cost of this check is measurable in the benchmarks we have for bloom filters:

```
name              old time/op    new time/op    delta
Fasthash            1.18ns ± 1%    1.19ns ± 1%     ~     (p=0.735 n=9+10)
BlockInsert         2.35ns ± 1%    2.25ns ± 1%   -4.39%  (p=0.000 n=9+10)
BlockCheck          2.10ns ± 2%    2.39ns ± 2%  +13.75%  (p=0.000 n=10+10)
FilterInsertBulk    16.9ns ± 2%    17.3ns ± 3%   +2.42%  (p=0.001 n=10+9)
FilterInsert        2.40ns ± 3%    2.77ns ±11%  +15.64%  (p=0.000 n=10+10)
FilterCheck         2.69ns ± 1%    2.97ns ± 1%  +10.67%  (p=0.000 n=10+8)

name              old speed      new speed      delta
Fasthash          27.0GB/s ± 1%  26.9GB/s ± 1%     ~     (p=0.661 n=9+10)
BlockInsert       13.6GB/s ± 1%  14.2GB/s ± 1%   +4.61%  (p=0.000 n=9+10)
BlockCheck        15.2GB/s ± 2%  13.4GB/s ± 2%  -12.10%  (p=0.000 n=10+10)
FilterInsertBulk  30.2GB/s ± 2%  29.5GB/s ± 3%   -2.36%  (p=0.001 n=10+9)
FilterInsert      13.3GB/s ± 3%  11.6GB/s ±10%  -13.33%  (p=0.000 n=10+10)
FilterCheck       11.9GB/s ± 1%  10.8GB/s ± 1%   -9.64%  (p=0.000 n=10+8)
```

The `Block*` and `FilterInsert` functions aren't very important, `FilterInsertBulk` and `FilterCheck` are the functions used by the parquet-go package.

* The `FilterInsertBulk` benchmark measures a 2% drop in throughput, this seems acceptable to me, especially considering we can amortize this further by increasing the number of values in the bulk insert to reduce this overhead
* The `FilterCheck` benchmark measures a ~10% drop in throughput, this function tests a single value hash in a bloom filter, so it is more sensitive to the new instructions. The check runs in <3ns, so it seems to be an acceptable as well.

It appears to have no measurable impact on the high level `SplitBloomFilter` benchmark:
```
name              old time/op    new time/op    delta
SplitBlockFilter    1.94µs ± 1%    1.92µs ± 0%  -0.72%  (p=0.001 n=10+8)

name              old speed      new speed      delta
SplitBlockFilter  4.13GB/s ± 1%  4.16GB/s ± 0%  +0.73%  (p=0.001 n=10+8)
```