@kevinburkesegment 2022-05-24
I didn't realize we could just leave CPU instructions in the binary and as long as they don't get executed by the CPU that can't execute them, we're OK. That's pretty cool.